AMBR-1075: s3-migrate: move common code to shared.py
